Get rid of misleading symmetry in StateUpdate.Combine
Hi Thomas, I just read your excellent post on the update monads. I stumbled for a few long minutes trying to understand why a \circ b of the `StateUpdate`s is so symmetrical unless both a and b are not noops, before I realized it is actually not! May I suggest a rewrite in this explicitly asymmetric form? To me it seems easier to understand: left to right, `SetNop` passes `v` through, but a `Set b` prevails.

Technically, the second case could have been written as `| _, v -> v`, but that would be didactically wrong, I think :)